+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # master
 
+- Add tests (@kiliankoe)
 - Added linting phase to danger (@BenchR267)
 - Add support for re-normalizing user input (@kiliankoe)
 - Added danger support (@BenchR267)